refactor: refactored the reducers into multiple files
