+++ client/store/index.js
@@ -2,14 +2,19 @@ import {createStore, combineReducers, applyMiddleware} from 'redux'
 import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
-import manageUsers from './reducer'
+import manageUsers from './user'
 import manageProducts from './products'
+import manageOrders from './orders'
 
-const reducer = combineReducers({user: manageUsers, products: manageProducts})
+const reducer = combineReducers({
+  user: manageUsers,
+  products: manageProducts,
+  orders: manageOrders
+})
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
 const store = createStore(reducer, middleware)
 
 export default store
-export * from './reducer'
+export * from './user'
