+++ client/store/user.js
@@ -1,29 +1,22 @@
 import axios from 'axios'
 import history from '../history'
-// import {combineReducers} from 'redux'
 
-/**
- * ACTION TYPES
- */
+// ACTION TYPES
+
 const GET_USER = 'GET_USER'
 const REMOVE_USER = 'REMOVE_USER'
-const CREATE_ORDER = 'CREATE_ORDER'
 
-/**
- * INITIAL STATE
- */
+// INITIAL STATE
+
 const defaultUser = {}
 
-/**
- * ACTION CREATORS
- */
+// ACTION CREATORS
+
 const getUser = user => ({type: GET_USER, user})
 const removeUser = () => ({type: REMOVE_USER})
-const createOrder = order => ({type: CREATE_ORDER, order})
 
-/**
- * THUNK CREATORS
- */
+// THUNK CREATORS
+
 export const me = () => async dispatch => {
   try {
     const res = await axios.get('/auth/me')
@@ -59,16 +52,8 @@ export const logout = () => async dispatch => {
   }
 }
 
-export const purchase = order => {
-  return async dispatch => {
-    const data = await Axios.post('/api/cart', order)
-    dispatch(createOrder(data))
-  }
-}
+// REDUCER
 
-/**
- * REDUCER
- */
 const manageUsers = (state = defaultUser, action) => {
   switch (action.type) {
     case GET_USER:
@@ -80,18 +65,4 @@ const manageUsers = (state = defaultUser, action) => {
   }
 }
 
-const manageOrder = (state = [], action) => {
-  switch (action.type) {
-    case CREATE_ORDER:
-      return action.order
-    default:
-      return state
-  }
-}
-
-// const reducer = combineReducers({
-//   user: manageUsers,
-//   products: manageProducts,
-//   order: manageOrder
-// })
 export default manageUsers
